engines/requirements.txt
google-cloud-bigquery
firebase-admin
requests
pandas
google-genai
google-generativeai
